Authenticate Firebase user via email and password
If the user did not previously exist in Firebase when the Signup/Login button is tapped, the user is created if it is valid; otherwise, any error condition is displayed via an alert.

Also, once the user has successfully logged in for the first time using email/password login, the app will automatically log them in each time the app is started so that they do not have to keep logging in.